chore: extract out isValidAllowedChildren
