+++ packages/blade/src/utils/index.ts
@@ -5,6 +5,7 @@ export * from './getIn';
 export * from './getPlatformType';
 export * from './isEmpty';
 export * from './isEqual';
+export * from './isValidAllowedChildren';
 export * from './isPartialMatchObjectKeys';
 export * from './makeAccessible';
 export * from './makeBezier';
@@ -20,5 +21,4 @@ export * from './toTitleCase';
 export * from './useBreakpoint';
 export * from './useColorScheme';
 export * from './usePrevious';
-export * from './isEmpty';
 export * from './platform';
