+++ packages/blade/src/utils/isValidAllowedChildren/index.ts
@@ -1,0 +1,1 @@
+export * from './isValidAllowedChildren';
